conflict's fixes
